On utilise directement Simulator dans City, Simulators n'étant qu'un wrapper inutile
