--- src/components/Coproprietaire/Simulators.tsx
@@ -1,66 +1,0 @@
-import dynamic from 'next/dynamic';
-
-import { ArrowItem } from '@/components/MarkdownWrapper/MarkdownWrapper.style';
-import Box, { ResponsiveRow } from '@/components/ui/Box';
-import Heading from '@/components/ui/Heading';
-import Link from '@/components/ui/Link';
-import Text from '@/components/ui/Text';
-
-const Simulator = dynamic(() => import('@/components/Ressources/Contents/Simulator'), {
-  ssr: false,
-});
-
-const Simulators = ({ textTitle, simulatorTitle }: { textTitle?: string; simulatorTitle?: string }) => {
-  return (
-    <Box className="fr-container">
-      <ResponsiveRow>
-        <Box flex>
-          <Heading as="h4" legacyColor="white">
-            {textTitle}
-          </Heading>
-          <Box>
-            <ArrowItem color="white">
-              <Text>
-                Le coup de pouce <strong>«&nbsp;Chauffage des bâtiments résidentiels collectifs et tertiaires&nbsp;»</strong> permet
-                d’obtenir des aides financières conséquentes pour se raccorder.
-              </Text>
-            </ArrowItem>
-            <ArrowItem color="white">
-              <Text>
-                <strong style={{ backgroundColor: '#F8D86E', color: '#000091' }}>
-                  Le coût du raccordement peut ainsi être réduit à quelques centaines d’euros par logement
-                </strong>{' '}
-                (en fonction de la situation du bâtiment et de ses besoins en chaleur).
-              </Text>
-            </ArrowItem>
-            <ArrowItem color="white">
-              <Text>
-                Différentes entreprises signataires de la charte « Chauffage des bâtiments résidentiels collectifs et tertiaires » offrent
-                cette prime. <br />
-                <strong>
-                  Le montant de la prime peut significativement varier d’une entreprise à l’autre, il est donc important de comparer les
-                  offres proposées.
-                </strong>
-              </Text>
-            </ArrowItem>
-          </Box>
-          <Box ml="4w" mt="1w">
-            <Link href="/ressources/aides#contenu" variant="primary">
-              Tout savoir sur cette aide
-            </Link>
-          </Box>
-        </Box>
-        <Box flex maxWidth="40%">
-          {simulatorTitle && (
-            <Heading as="h4" legacyColor="white">
-              {simulatorTitle}
-            </Heading>
-          )}
-          <Simulator />
-        </Box>
-      </ResponsiveRow>
-    </Box>
-  );
-};
-
-export default Simulators;
